// src/app/models/event.model.ts
export interface Event {
    event_id: string;
    event_name: string;
    event_date: string;
    event_time: string;
    location: string;
    capacity: number;
    per_person_price: number;
    event_description: string;
    event_type:string;
    speaker_name: string;
    organizer_name: string;
    created_at: string;
  }